Include pending transactions in Client#get_nonce
Reason: Submitting more than one transaction from the same account
before the block is mined resulted with colliding transactions, because
of the same nonce. This change takes the nonce from the pending
transactions, so it's incrementes as the pool growths before block is
mined.

Relates to: https://github.com/EthWorks/ethereum.rb/issues/35